refactor(routers): move web presentation routes
